changed return type of the root function

+++ 8_square_root_of_a_number.c
@@ -9,7 +9,7 @@ int input_num()
     return num;
 }
 
-int square_root(int num)
+float square_root(int num)
 {
     return(sqrt(num));
 }
